cleanup: fix warnings from clang/gcc -Weverything+-fanalyzer, cppcheck, clang-tidy
Compiled clean (no warnings) under system clang, Homebrew LLVM clang, and
gcc-15 with -Wall -Wextra plus -Weverything/-fanalyzer; ran cppcheck and
clang-tidy over the tree. Remaining findings in all four tools were either
false positives (STL-container exception-unwind speculation in the static
analyzers, signed-shift/memset patterns that are standard-compliant and
required for bit-exactness) or purely stylistic (old-style-cast,
sign-conversion, struct padding) inherent to this decompiled/RE-derived
codebase, left as-is.

Real fixes:
- explicit casts documenting intentional narrowing (64->32 bit, double->int,
  byte-packing) instead of relying on implicit conversions
- OFR_SolveLDLT / decode_tree_symbol given internal linkage (only used
  within their own TU)
- explicit constructor on OFR_BitStream and OFR_PredictorDualStreamStereo's
  sibling issue on optimfrogcpp::DecodeError's weak vtable (out-of-line dtor)
- value-initialize a few locals cppcheck/gcc flagged as possibly read
  uninitialized (behaviorally inert, just silences the warning)
- defensive clamp on encoder tuning indices read from env vars
- all three CLI main()s now wrap their body in try/catch instead of letting
  exceptions escape into std::terminate()
- added Doxygen comments to include/optimfrogcpp.hpp, the one hand-written
  C++-namespaced header in the project

Verified: conformance suite still PASS=1760 FAIL=0 (no regressions), and a
clean ASan+UBSan round-trip (encode/decode, lossless + Float + DualStream,
mono/stereo) with zero sanitizer reports.

+++ src/core/optimfrog_encoder_float.cpp
@@ -25,7 +25,9 @@ static uint32_t crc32_calc_f(const uint8_t* p, size_t n) {
     for (size_t i = 0; i < n; i++) c = crc32_table_f[(c ^ p[i]) & 0xff] ^ (c >> 8);
     return c ^ 0xFFFFFFFFu;
 }
-static void put32f(std::vector<uint8_t>& o, uint32_t v) { o.push_back(v); o.push_back(v>>8); o.push_back(v>>16); o.push_back(v>>24); }
+static void put32f(std::vector<uint8_t>& o, uint32_t v) {
+    o.push_back((uint8_t)v); o.push_back((uint8_t)(v>>8)); o.push_back((uint8_t)(v>>16)); o.push_back((uint8_t)(v>>24));
+}
 
 struct FloatBlockStats { int32_t minExp, maxExp; };
